+++ decomp/ST.exe/functions/0075BA30/decomp.c
@@ -26,7 +26,7 @@ undefined4 FUN_0075ba30(AnonShape_0075BA30_FEA40A39 *param_1,undefined4 *param_2
   int local_10;
   int *local_c;
   undefined4 *local_8;
-  
+
   iVar7 = param_1->field_01AA;
   local_20 = iVar7;
   if (((param_1->field_010C != 0) && (*(int *)(iVar7 + 0x24) == 0)) &&
@@ -56,7 +56,7 @@ undefined4 FUN_0075ba30(AnonShape_0075BA30_FEA40A39 *param_1,undefined4 *param_2
       local_10 = *(int *)(iVar7 + 0x28 + *(int *)(iVar4 + 0x14) * 4);
       iVar7 = *(int *)(iVar7 + 0x38 + *(int *)(iVar4 + 0x18) * 4);
       if (iVar2 < 8) {
-        iVar2 = FUN_0075b720(&local_50,(uint)pAVar8,iVar2,0);
+        iVar2 = FUN_0075b720(&local_50,pAVar8,iVar2,0);
         if (iVar2 == 0) {
           return 0;
         }
@@ -75,6 +75,7 @@ LAB_0075bb5c:
       else {
 LAB_0075bb2d:
         uVar3 = (int)pAVar8 >> ((char)iVar2 - 8U & 0x1f) & 0xff;
+        /* ST_PSEUDO[raw_pointer_offset]: candidate structure field after proof; otherwise retain buffer arithmetic */
         iVar5 = *(int *)(local_10 + 0xd4 + uVar3 * 4);
         if (iVar5 == 0) {
           iVar5 = 9;
@@ -86,8 +87,8 @@ LAB_0075bb2d:
       uVar6 = 0;
       if (uVar3 != 0) {
         if ((iVar2 < (int)uVar3) &&
-           (iVar4 = FUN_0075b720(&local_50,(uint)pAVar8,iVar2,uVar3), iVar2 = local_40,
-           pAVar8 = local_44, iVar4 == 0)) {
+           (iVar4 = FUN_0075b720(&local_50,pAVar8,iVar2,uVar3), iVar2 = local_40, pAVar8 = local_44,
+           iVar4 == 0)) {
           return 0;
         }
         iVar2 = iVar2 - uVar3;
@@ -103,7 +104,7 @@ LAB_0075bcfa:
         iVar4 = 1;
         do {
           if (iVar2 < 8) {
-            iVar2 = FUN_0075b720(&local_50,(uint)pAVar8,iVar2,0);
+            iVar2 = FUN_0075b720(&local_50,pAVar8,iVar2,0);
             if (iVar2 == 0) {
               return 0;
             }
@@ -136,7 +137,7 @@ LAB_0075bd2e:
           }
           else {
             if ((iVar2 < (int)uVar6) &&
-               (iVar5 = FUN_0075b720(&local_50,(uint)pAVar8,iVar2,uVar6), iVar2 = local_40,
+               (iVar5 = FUN_0075b720(&local_50,pAVar8,iVar2,uVar6), iVar2 = local_40,
                pAVar8 = local_44, iVar5 == 0)) {
               return 0;
             }
@@ -153,7 +154,7 @@ LAB_0075bd2e:
         iVar4 = 1;
         do {
           if (iVar2 < 8) {
-            iVar2 = FUN_0075b720(&local_50,(uint)pAVar8,iVar2,0);
+            iVar2 = FUN_0075b720(&local_50,pAVar8,iVar2,0);
             if (iVar2 == 0) {
               return 0;
             }
@@ -188,7 +189,7 @@ LAB_0075bc29:
           else {
             local_24 = iVar4 + ((int)uVar3 >> 4);
             if ((iVar2 < (int)uVar6) &&
-               (iVar4 = FUN_0075b720(&local_50,(uint)pAVar8,iVar2,uVar6), iVar2 = local_40,
+               (iVar4 = FUN_0075b720(&local_50,pAVar8,iVar2,uVar6), iVar2 = local_40,
                pAVar8 = local_44, iVar4 == 0)) {
               return 0;
             }
